拆分 webui 的 oneday 部分
